Charting Update - Combined Data, Edges, Regulogs
Updated the charting process to combine the binding site and
regulated gene information. Weighting of genes is now based on
the number of binding site matches.

Locus Tags are used as the matching mechanism.

The edge thickness is now indicative of the number of matching
binding sites.

The binding sites are now shown on leaf tooltips.

Fixed the bug where Regulogs weren't being displayed correctly.